Feature/dashboard 3 (#42)
* style(product-page): fix styling issues on product page

Correct styling inconsistencies on the product page, including adjustments to layout, spacing, and typography for a more polished and consistent user experience.

* feat(product): display stock on card and product description, disable button when stock is 0
Add stock information to the product card and product description page to improve user awareness of availability. Also, disable the Add to Cart button when stock is 0 to prevent users from attempting to purchase out-of-stock items.

* feat(landing-page): make landing page responsive and connect to product database

Ensure the landing page is fully responsive across different devices. Connect the landing page to the product database to dynamically display featured products. Implement navigation to the product page based on category when See All is clicked, directing users to the relevant product category.

* feat(product-detail): add 'Picks for You' and 'Explore More Products' carousels below product details to suggest related products and an to showcase additional items. Both carousels are placed below the product details to enhance user engagement and browsing experience.

* feat(product): add stock management feature

* feat(stock): add crud stock

* fix(stock): add crud stock

* feat(stock): fix admin page fetch warehouse select

* feat(stock report): add crud stock report andstock mutation

* feat: create report and dashboard page for admin

* fix: conflict report and dashboard v3

* fix: conflict report and dashboard v4

* fix: conflict report and dashboard v6

* fix: conflict report and dashboard v7

* fix: conflict report and dashboard v7

* fix: conflict report and dashboard v7

* fix: conflict report and dashboard v7

* fix: conflict report and dashboard v7

* fix: conflict report and dashboard v7

* fix: conflict report and dashboard v7